Reorganize Client Setup functions
