+++ modules/server/src/index.js
@@ -1,4 +1,4 @@
 export { default as App } from './app.js';
 export { createSchemasFromConfigs, default as getGraphQLRoutes } from './graphqlRoutes.js';
 export { default } from './server.js';
-export { default as getSearchClient } from './searchClient/index.js';
+export { default as getSearchClient, buildSearchClient } from './searchClient/index.js';
